🔨Refactor: 비밀번호 재설정 페이지 파일 구조 수정
