--- src/pages/password-update/PasswordUpdate.styles.tsx
@@ -1,52 +1,0 @@
-import styled from '@emotion/styled';
-
-const LogoContainer = styled.div`
-  margin-bottom: 80px;
-`;
-
-const Logo = () => {
-  return (
-    <svg
-      width='223'
-      height='44'
-      viewBox='0 0 107 14'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'>
-      <path
-        d='M0.0639063 -9.53674e-07H5.94391L11.7839 7V-9.53674e-07H12.9839V14H8.28391L1.28391 4.68V14H0.0639063V-9.53674e-07ZM14.1069 -9.53674e-07H20.6269V14H14.1069V-9.53674e-07ZM21.7631 -9.53674e-07H35.1431C35.7965 -9.53674e-07 36.3965 0.166666 36.9431 0.5C37.5031 0.82 37.9431 1.26 38.2631 1.82C38.5965 2.36667 38.7631 2.96667 38.7631 3.62C38.7631 4.48667 38.4965 5.27333 37.9631 5.98C37.4298 6.68667 36.7565 7.14 35.9431 7.34C36.6631 7.54 37.2498 7.92 37.7031 8.48C38.1698 9.04 38.4031 9.67333 38.4031 10.38V14H31.7031V9.66C31.7031 9.23333 31.6231 8.9 31.4631 8.66C31.2765 8.38 31.0631 8.17333 30.8231 8.04C30.5565 7.89333 30.2565 7.78667 29.9231 7.72C29.5765 7.65333 29.2498 7.61333 28.9431 7.6L28.4031 7.58V14H21.7631V-9.53674e-07ZM29.1031 6.42C29.5831 6.42 30.0231 6.30667 30.4231 6.08C30.8365 5.84 31.1565 5.52 31.3831 5.12C31.6231 4.70667 31.7431 4.26667 31.7431 3.8C31.7431 3.32 31.6231 2.88 31.3831 2.48C31.1565 2.06667 30.8365 1.74 30.4231 1.5C30.0231 1.26 29.5831 1.14 29.1031 1.14H28.4031V6.44L29.1031 6.42ZM39.9309 -9.53674e-07H53.2709C53.9509 -9.53674e-07 54.5709 0.166666 55.1309 0.5C55.7043 0.82 56.1576 1.26667 56.4909 1.84C56.8243 2.4 56.9909 3.02 56.9909 3.7C56.9909 4.35333 56.8309 4.97333 56.5109 5.56C56.1909 6.14667 55.7643 6.62 55.2309 6.98C55.7643 7.35333 56.1909 7.84 56.5109 8.44C56.8309 9.02667 56.9909 9.64667 56.9909 10.3C56.9909 10.9667 56.8243 11.5867 56.4909 12.16C56.1576 12.72 55.7043 13.1667 55.1309 13.5C54.5709 13.8333 53.9509 14 53.2709 14H39.9309V-9.53674e-07ZM47.2509 6.42C47.7309 6.42 48.1709 6.30667 48.5709 6.08C48.9843 5.84 49.3043 5.52 49.5309 5.12C49.7709 4.70667 49.8909 4.26667 49.8909 3.8C49.8909 3.32 49.7709 2.88 49.5309 2.48C49.3043 2.06667 48.9843 1.74 48.5709 1.5C48.1709 1.26 47.7309 1.14 47.2509 1.14H46.9509V6.4C47.0176 6.41333 47.1176 6.42 47.2509 6.42ZM47.2509 12.84C47.7309 12.84 48.1709 12.7267 48.5709 12.5C48.9843 12.26 49.3043 11.94 49.5309 11.54C49.7709 11.1267 49.8909 10.68 49.8909 10.2C49.8909 9.73333 49.7709 9.3 49.5309 8.9C49.3043 8.48667 48.9843 8.16667 48.5709 7.94C48.1709 7.7 47.7309 7.58 47.2509 7.58H46.9509V12.82C47.0176 12.8333 47.1176 12.84 47.2509 12.84ZM62.4095 -9.53674e-07H68.2495L74.4495 14H66.8895L64.7895 9.36H60.3895L58.5895 14H57.1295L62.4095 -9.53674e-07ZM64.2295 8.14L62.4095 4.08L60.8295 8.14H64.2295ZM75.2397 -9.53674e-07H81.1197L86.9597 7V-9.53674e-07H88.1597V14H83.4597L76.4597 4.68V14H75.2397V-9.53674e-07ZM94.0892 -9.53674e-07H99.9292L106.129 14H98.5692L96.4692 9.36H92.0692L90.2692 14H88.8092L94.0892 -9.53674e-07ZM95.9092 8.14L94.0892 4.08L92.5092 8.14H95.9092Z'
-        fill='white'
-      />
-    </svg>
-  );
-};
-
-const PasswordUpdateContainer = styled.div`
-  ${({ theme }) => theme.style.flexCenter};
-  flex-direction: column;
-  width: 100vw;
-  height: calc(100vh - 64px);
-  background: ${({ theme }) => theme.color.linearGradientPurple};
-`;
-
-const PasswordUpdateForm = styled.form`
-  background-color: ${({ theme }) => theme.color.white};
-  ${({ theme }) => theme.style.flexCenter};
-  flex-direction: column;
-  width: 330px;
-  height: 390px;
-  border-radius: 10px;
-`;
-
-const ButtonContainer = styled.div`
-  ${({ theme }) => theme.style.flexJustifyCenter};
-  width: 300px;
-  margin-top: 15px;
-`;
-
-export {
-  PasswordUpdateForm,
-  ButtonContainer,
-  PasswordUpdateContainer,
-  LogoContainer,
-  Logo
-};
